refactor: wrap common function to utils
