+++ perception/probabilistic_occupancy_grid_map/src/pointcloud_based_occupancy_grid_map/occupancy_grid_map.cpp
@@ -52,6 +52,7 @@
 #include "pointcloud_based_occupancy_grid_map/occupancy_grid_map.hpp"
 
 #include "cost_value.hpp"
+#include "utils/utils.hpp"
 
 #include <pcl_ros/transforms.hpp>
 #include <tier4_autoware_utils/tier4_autoware_utils.hpp>
@@ -68,44 +69,6 @@
 #endif
 
 #include <algorithm>
-namespace
-{
-void transformPointcloud(
-  const sensor_msgs::msg::PointCloud2 & input, const geometry_msgs::msg::Pose & pose,
-  sensor_msgs::msg::PointCloud2 & output)
-{
-  const auto transform = tier4_autoware_utils::pose2transform(pose);
-  Eigen::Matrix4f tf_matrix = tf2::transformToEigen(transform).matrix().cast<float>();
-
-  pcl_ros::transformPointCloud(tf_matrix, input, output);
-  output.header.stamp = input.header.stamp;
-  output.header.frame_id = "";
-}
-
-/**
- * @brief Get the Inverse Pose object
- *
- * @param input
- * @return geometry_msgs::msg::Pose inverted pose
- */
-geometry_msgs::msg::Pose getInversePose(const geometry_msgs::msg::Pose & pose)
-{
-  tf2::Vector3 position(pose.position.x, pose.position.y, pose.position.z);
-  tf2::Quaternion orientation(
-    pose.orientation.x, pose.orientation.y, pose.orientation.z, pose.orientation.w);
-  tf2::Transform tf(orientation, position);
-  const auto inv_tf = tf.inverse();
-  geometry_msgs::msg::Pose inv_pose;
-  inv_pose.position.x = inv_tf.getOrigin().x();
-  inv_pose.position.y = inv_tf.getOrigin().y();
-  inv_pose.position.z = inv_tf.getOrigin().z();
-  inv_pose.orientation.x = inv_tf.getRotation().x();
-  inv_pose.orientation.y = inv_tf.getRotation().y();
-  inv_pose.orientation.z = inv_tf.getRotation().z();
-  inv_pose.orientation.w = inv_tf.getRotation().w();
-  return inv_pose;
-}
-}  // namespace
 
 namespace costmap_2d
 {
@@ -203,14 +166,14 @@ void OccupancyGridMap::updateWithPointCloud(
 
   // Transform from base_link to map frame
   PointCloud2 map_raw_pointcloud, map_obstacle_pointcloud;  // point cloud in map frame
-  transformPointcloud(raw_pointcloud, robot_pose, map_raw_pointcloud);
-  transformPointcloud(obstacle_pointcloud, robot_pose, map_obstacle_pointcloud);
+  utils::transformPointcloud(raw_pointcloud, robot_pose, map_raw_pointcloud);
+  utils::transformPointcloud(obstacle_pointcloud, robot_pose, map_obstacle_pointcloud);
 
   // Transform from map frame to scan frame
-  PointCloud2 scan_raw_pointcloud, scan_obstacle_pointcloud;  // point cloud in scan frame
-  const auto scan2map_pose = getInversePose(scan_origin);     // scan -> map transform pose
-  transformPointcloud(map_raw_pointcloud, scan2map_pose, scan_raw_pointcloud);
-  transformPointcloud(map_obstacle_pointcloud, scan2map_pose, scan_obstacle_pointcloud);
+  PointCloud2 scan_raw_pointcloud, scan_obstacle_pointcloud;      // point cloud in scan frame
+  const auto scan2map_pose = utils::getInversePose(scan_origin);  // scan -> map transform pose
+  utils::transformPointcloud(map_raw_pointcloud, scan2map_pose, scan_raw_pointcloud);
+  utils::transformPointcloud(map_obstacle_pointcloud, scan2map_pose, scan_obstacle_pointcloud);
 
   // Create angle bins
   struct BinInfo
