+++ perception/probabilistic_occupancy_grid_map/src/pointcloud_based_occupancy_grid_map/pointcloud_based_occupancy_grid_map_node.cpp
@@ -15,6 +15,7 @@
 #include "pointcloud_based_occupancy_grid_map/pointcloud_based_occupancy_grid_map_node.hpp"
 
 #include "cost_value.hpp"
+#include "utils/utils.hpp"
 
 #include <pcl_ros/transforms.hpp>
 #include <tier4_autoware_utils/tier4_autoware_utils.hpp>
@@ -35,164 +36,6 @@
 #include <algorithm>
 #include <memory>
 #include <string>
-
-namespace
-{
-bool transformPointcloud(
-  const sensor_msgs::msg::PointCloud2 & input, const tf2_ros::Buffer & tf2,
-  const std::string & target_frame, sensor_msgs::msg::PointCloud2 & output)
-{
-  geometry_msgs::msg::TransformStamped tf_stamped;
-  tf_stamped = tf2.lookupTransform(
-    target_frame, input.header.frame_id, input.header.stamp, rclcpp::Duration::from_seconds(0.5));
-  // transform pointcloud
-  Eigen::Matrix4f tf_matrix = tf2::transformToEigen(tf_stamped.transform).matrix().cast<float>();
-  pcl_ros::transformPointCloud(tf_matrix, input, output);
-  output.header.stamp = input.header.stamp;
-  output.header.frame_id = target_frame;
-  return true;
-}
-
-bool cropPointcloudByHeight(
-  const sensor_msgs::msg::PointCloud2 & input, const tf2_ros::Buffer & tf2,
-  const std::string & target_frame, const float min_height, const float max_height,
-  sensor_msgs::msg::PointCloud2 & output)
-{
-  rclcpp::Clock clock{RCL_ROS_TIME};
-  // Transformed pointcloud on target frame
-  sensor_msgs::msg::PointCloud2 trans_input_tmp;
-  const bool is_target_frame = (input.header.frame_id == target_frame);
-  if (!is_target_frame) {
-    if (!transformPointcloud(input, tf2, target_frame, trans_input_tmp)) return false;
-  }
-  const sensor_msgs::msg::PointCloud2 & trans_input = is_target_frame ? input : trans_input_tmp;
-
-  // Apply height filter
-  pcl::PointCloud<pcl::PointXYZ>::Ptr pcl_output(new pcl::PointCloud<pcl::PointXYZ>);
-  for (sensor_msgs::PointCloud2ConstIterator<float> iter_x(trans_input, "x"),
-       iter_y(trans_input, "y"), iter_z(trans_input, "z");
-       iter_x != iter_x.end(); ++iter_x, ++iter_y, ++iter_z) {
-    if (min_height < *iter_z && *iter_z < max_height) {
-      pcl_output->push_back(pcl::PointXYZ(*iter_x, *iter_y, *iter_z));
-    }
-  }
-
-  // Convert to ros msg
-  pcl::toROSMsg(*pcl_output, output);
-  output.header = trans_input.header;
-  return true;
-}
-
-geometry_msgs::msg::Pose getPose(
-  const std_msgs::msg::Header & source_header, const tf2_ros::Buffer & tf2,
-  const std::string & target_frame)
-{
-  geometry_msgs::msg::Pose pose;
-  geometry_msgs::msg::TransformStamped tf_stamped;
-  tf_stamped = tf2.lookupTransform(
-    target_frame, source_header.frame_id, source_header.stamp, rclcpp::Duration::from_seconds(0.5));
-  pose = tier4_autoware_utils::transform2pose(tf_stamped.transform);
-  return pose;
-}
-
-geometry_msgs::msg::Pose getPose(
-  const builtin_interfaces::msg::Time & stamp, const tf2_ros::Buffer & tf2,
-  const std::string & source_frame, const std::string & target_frame)
-{
-  geometry_msgs::msg::Pose pose;
-  geometry_msgs::msg::TransformStamped tf_stamped;
-  tf_stamped =
-    tf2.lookupTransform(target_frame, source_frame, stamp, rclcpp::Duration::from_seconds(0.5));
-  pose = tier4_autoware_utils::transform2pose(tf_stamped.transform);
-  return pose;
-}
-
-/**
- * @brief 3D point struct for sorting and searching
- *
- */
-struct MyPoint3d
-{
-  float x;
-  float y;
-  float z;
-
-  // constructor
-  MyPoint3d(float x, float y, float z) : x(x), y(y), z(z) {}
-
-  // calc squared norm
-  float norm2() const { return powf(x, 2) + powf(y, 2) + powf(z, 2); }
-
-  // calc arctan2
-  float arctan2() const { return atan2f(y, x); }
-
-  // overload operator<
-  bool operator<(const MyPoint3d & other) const
-  {
-    const auto a = norm2();
-    const auto b = other.norm2();
-    if (a == b) {  // escape when norm2 is same
-      return arctan2() < other.arctan2();
-    } else {
-      return a < b;
-    }
-  }
-
-  // overload operator==
-  bool operator==(const MyPoint3d & other) const
-  {
-    return fabsf(x - other.x) < FLT_EPSILON && fabsf(y - other.y) < FLT_EPSILON &&
-           fabsf(z - other.z) < FLT_EPSILON;
-  }
-};
-
-/**
- * @brief extract Common Pointcloud between obstacle pc and raw pc
- * @param obstacle_pc
- * @param raw_pc
- * @param output_obstacle_pc
- */
-bool extractCommonPointCloud(
-  const sensor_msgs::msg::PointCloud2 & obstacle_pc, const sensor_msgs::msg::PointCloud2 & raw_pc,
-  sensor_msgs::msg::PointCloud2 & output_obstacle_pc)
-{
-  // Convert to vector of 3d points
-  std::vector<MyPoint3d> v_obstacle_pc, v_raw_pc, v_output_obstacle_pc;
-  for (sensor_msgs::PointCloud2ConstIterator<float> iter_x(obstacle_pc, "x"),
-       iter_y(obstacle_pc, "y"), iter_z(obstacle_pc, "z");
-       iter_x != iter_x.end(); ++iter_x, ++iter_y, ++iter_z) {
-    v_obstacle_pc.push_back(MyPoint3d(*iter_x, *iter_y, *iter_z));
-  }
-  for (sensor_msgs::PointCloud2ConstIterator<float> iter_x(raw_pc, "x"), iter_y(raw_pc, "y"),
-       iter_z(raw_pc, "z");
-       iter_x != iter_x.end(); ++iter_x, ++iter_y, ++iter_z) {
-    v_raw_pc.push_back(MyPoint3d(*iter_x, *iter_y, *iter_z));
-  }
-
-  // sort pointclouds for searching cross points: O(nlogn)
-  std::sort(v_obstacle_pc.begin(), v_obstacle_pc.end(), [](auto a, auto b) { return a < b; });
-  std::sort(v_raw_pc.begin(), v_raw_pc.end(), [](auto a, auto b) { return a < b; });
-
-  // calc intersection points of two pointclouds: O(n)
-  set_intersection(
-    v_obstacle_pc.begin(), v_obstacle_pc.end(), v_raw_pc.begin(), v_raw_pc.end(),
-    std::back_inserter(v_output_obstacle_pc));
-  if (v_output_obstacle_pc.size() == 0) {
-    return false;
-  }
-
-  // Convert to ros msg
-  pcl::PointCloud<pcl::PointXYZ>::Ptr pcl_output(new pcl::PointCloud<pcl::PointXYZ>);
-  for (auto p : v_output_obstacle_pc) {
-    pcl_output->push_back(pcl::PointXYZ(p.x, p.y, p.z));
-  }
-  pcl::toROSMsg(*pcl_output, output_obstacle_pc);
-  output_obstacle_pc.header = obstacle_pc.header;
-
-  return true;
-}
-
-}  // namespace
 
 namespace occupancy_grid_map
 {
@@ -259,12 +102,12 @@ void PointcloudBasedOccupancyGridMapNode::onPointcloudWithObstacleAndRaw(
   PointCloud2 cropped_raw_pc{};
   if (use_height_filter_) {
     constexpr float min_height = -1.0, max_height = 2.0;
-    if (!cropPointcloudByHeight(
+    if (!utils::cropPointcloudByHeight(
           *input_obstacle_msg, *tf2_, base_link_frame_, min_height, max_height,
           cropped_obstacle_pc)) {
       return;
     }
-    if (!cropPointcloudByHeight(
+    if (!utils::cropPointcloudByHeight(
           *input_raw_msg, *tf2_, base_link_frame_, min_height, max_height, cropped_raw_pc)) {
       return;
     }
@@ -276,7 +119,7 @@ void PointcloudBasedOccupancyGridMapNode::onPointcloudWithObstacleAndRaw(
   // Filter obstacle pointcloud by raw pointcloud
   PointCloud2 filtered_obstacle_pc_common{};
   if (filter_obstacle_pointcloud_by_raw_pointcloud_) {
-    if (!extractCommonPointCloud(
+    if (!utils::extractCommonPointCloud(
           filtered_obstacle_pc, filtered_raw_pc, filtered_obstacle_pc_common)) {
       filtered_obstacle_pc_common = filtered_obstacle_pc;
     }
@@ -289,9 +132,11 @@ void PointcloudBasedOccupancyGridMapNode::onPointcloudWithObstacleAndRaw(
   Pose gridmap_origin{};
   Pose scan_origin{};
   try {
-    robot_pose = getPose(input_raw_msg->header, *tf2_, map_frame_);
-    gridmap_origin = getPose(input_raw_msg->header.stamp, *tf2_, gridmap_origin_frame_, map_frame_);
-    scan_origin = getPose(input_raw_msg->header.stamp, *tf2_, scan_origin_frame_, map_frame_);
+    robot_pose = utils::getPose(input_raw_msg->header, *tf2_, map_frame_);
+    gridmap_origin =
+      utils::getPose(input_raw_msg->header.stamp, *tf2_, gridmap_origin_frame_, map_frame_);
+    scan_origin =
+      utils::getPose(input_raw_msg->header.stamp, *tf2_, scan_origin_frame_, map_frame_);
   } catch (tf2::TransformException & ex) {
     RCLCPP_WARN_STREAM(get_logger(), ex.what());
     return;
